pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "long-parser"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "python-docx>=1.0",
]

[project.optional-dependencies]
retrieve = [
    "qdrant-client>=1.7",
    "sentence-transformers>=2.2",
]
web = [
    "fastapi>=0.110",
    "uvicorn[standard]>=0.27",
    "python-multipart>=0.0.9",
    "jinja2>=3.1",
    "python-docx>=1.0",
    "qdrant-client>=1.7",
    "sentence-transformers>=2.2",
]

[tool.setuptools.packages.find]
where = ["src"]
